Database.EventStore

Description

Connection

data Connection

Represents a connection to a single EventStore node.

data ConnectionType

Gathers every connection type handled by the client.

Constructors

Static String Int
HostName and Port.

Cluster ClusterSettings

Dns ByteString (Maybe DnsServer) Int
Domain name, optional DNS server and port.

data Credentials

Holds login and password information.

data Settings

Global Connection settings

Constructors

Settings

Fields

s_heartbeatInterval :: NominalDiffTime

Maximum delay of inactivity before the client sends a heartbeat request.

s_heartbeatTimeout :: NominalDiffTime

Maximum delay the server has to issue a heartbeat response.

s_requireMaster :: Bool

On a cluster settings. Requires the master node when performing a write operation.

s_retry :: Retry

Retry strategy when failing to connect.

s_reconnect_delay :: NominalDiffTime

Delay before issuing a new connection request.

s_ssl :: Maybe TLSSettings

SSL settings.

s_loggerType :: LogType

Type of logging to use.

s_loggerFilter :: LoggerFilter

Restriction of what would be logged.

s_loggerDetailed :: Bool

Detailed logging output. Currently, it also indicates the location where the log occurred.

s_operationTimeout :: NominalDiffTime

Delay in which an operation will be retried if no response arrived.

s_operationRetry :: Retry

Retry strategy when an operation timeout.

s_monitoring :: MonitoringBackend

Monitoring backend abstraction. You could implement one targetting `ekg-core` for example. We will expose an `ekg-core` implementation as soon as `ekg-core` supports GHC 8.8.*.

s_defaultConnectionName :: Maybe Text

Default connection name.

s_defaultUserCredentials :: Maybe Credentials

Credentials to use for operations where other Credentials are not explicitly supplied.

data Retry

Represents reconnection strategy.

atMost :: Int -> Retry

Indicates how many times we should try to reconnect to the server. A value less than or equal to 0 means no retry.

keepRetrying :: Retry

Indicates we should try to reconnect to the server until the end of the Universe.

credentials

Arguments

:: ByteString
Login

-> ByteString
Password

-> Credentials

Creates a Credentials given a login and a password.

defaultSettings :: Settings

Default global settings.

s_heartbeatInterval = 750 ms

s_heartbeatTimeout = 1500 ms

s_requireMaster = True

s_retry = atMost 3

s_reconnect_delay = 3 seconds

s_ssl = Nothing

s_loggerType = LogNone

s_loggerFilter = LoggerLevel LevelInfo

s_loggerDetailed = False

s_operationTimeout = 10 seconds

s_operationRetry = atMost 3

s_monitoring = noopMonitoringBackend

s_defaultConnectionName = Nothing

s_defaultUserCredentials = Nothing

defaultSSLSettings :: TLSSettings -> Settings

Default SSL settings based on defaultSettings.

connect :: Settings -> ConnectionType -> IO Connection

Creates a new Connection to a single node. It maintains a full duplex connection to the EventStore. An EventStore Connection operates quite differently than say a SQL connection. Normally when you use an EventStore connection you want to keep the connection open for a much longer of time than when you use a SQL connection.

Another difference is that with the EventStore Connection all operations are handled in a full async manner (even if you call the synchronous behaviors). Many threads can use an EvenStore Connection at the same time or a single thread can make many asynchronous requests. To get the most performance out of the connection it is generally recommended to use it in this way.

shutdown :: Connection -> IO ()

Asynchronously closes the Connection.

waitTillClosed :: Connection -> IO ()

Waits the Connection to be closed.

connectionSettings :: Connection -> Settings

Returns a Connection 's Settings.

Cluster Connection

data ClusterSettings

Contains settings related to a connection to a cluster.

Constructors

ClusterSettings

Fields

clusterDns :: !ByteString

The DNS name to use for discovering endpoints.

clusterMaxDiscoverAttempts :: !Int

The maximum number of attempts for discovering endpoints.

clusterExternalGossipPort :: !Int

The well-known endpoint on which cluster managers are running.

clusterGossipSeeds :: Maybe (NonEmpty GossipSeed)

Endpoints for seeding gossip if not using DNS.

clusterGossipTimeout :: !TimeSpan

Timeout for cluster gossip.

clusterDnsServer :: !(Maybe DnsServer)

Indicates a specific DNS server should be contacted.

data DnsServer

Tells how the DNS server should be contacted.

Constructors

DnsFilePath String

DnsHostName String

DnsHostPort String Int

data GossipSeed

Represents a source of cluster gossip.

gossipSeed :: String -> Int -> GossipSeed

Creates a GossipSeed.

gossipSeedWithHeader :: String -> Int -> String -> GossipSeed

Creates a GossipSeed with a specific HTTP header.

gossipSeedHost :: GossipSeed -> String

Returns GossipSeed host IP address.

gossipSeedHeader :: GossipSeed -> String

The host header to be sent when requesting gossip.

gossipSeedPort :: GossipSeed -> Int

Returns GossipSeed port.

gossipSeedClusterSettings :: NonEmpty GossipSeed -> ClusterSettings

Configures a ClusterSettings for connecting to a cluster using gossip seeds. clusterDns = "" clusterMaxDiscoverAttempts = 10 clusterExternalGossipPort = 0 clusterGossipTimeout = 1s

dnsClusterSettings :: ByteString -> ClusterSettings

Configures a ClusterSettings for connecting to a cluster using DNS discovery. clusterMaxDiscoverAttempts = 10 clusterExternalGossipPort = 0 clusterGossipSeeds = Nothing clusterGossipTimeout = 1s

Event

data Event

Contains event information like its type and data. Only used for write queries.

data EventData

Holds event data.

data EventType

Constants for System event types.

Constructors

StreamDeletedType
Event type for stream deleted.

StatsCollectionType
Event type for statistics.

LinkToType
Event type for linkTo.

StreamMetadataType
Event type for stream metadata.

SettingsType
Event type for the system settings.

UserDefined Text
Event defined by the user.

createEvent

Arguments

:: EventType
Event type

-> Maybe UUID
Event ID, generated if Nothing

-> EventData
Event data

-> Event

Create an Event meant to be persisted.

withJson :: ToJSON a => a -> EventData

Creates an event using JSON format

withJsonAndMetadata :: (ToJSON a, ToJSON b) => a -> b -> EventData

Creates an event with metadata using JSON format.

withBinary :: ByteString -> EventData

Creates an event using a binary format.

withBinaryAndMetadata :: ByteString -> ByteString -> EventData

Creates an event with metadata using binary format.

Event number

data EventNumber

Represents an event position within a stream.

streamStart :: EventNumber

The first event in a stream.

streamEnd :: EventNumber

The last event in the stream.

eventNumber :: Natural -> EventNumber

the Nth event of a stream.

rawEventNumber :: Int64 -> EventNumber

Returns a 'EventNumber from a raw Int64.

eventNumberToInt64 :: EventNumber -> Int64

Returns a raw Int64 from an EventNumber.

Read Operations

data StreamMetadataResult

Represents stream metadata as a series of properties for system data and a StreamMetadata object for user metadata.

Constructors

StreamMetadataResult

Fields

streamMetaResultStream :: !Text

The name of the stream.

streamMetaResultVersion :: !Int64

The version of the metadata format.

streamMetaResultData :: !StreamMetadata

A StreamMetadata containing user-specified metadata.

NotFoundStreamMetadataResult
When the stream is either not found or 'no stream'.

Fields

streamMetaResultStream :: !Text

The name of the stream.

DeletedStreamMetadataResult
When the stream is soft-deleted.

Fields

streamMetaResultStream :: !Text

The name of the stream.

type family BatchResult t where ...

When batch-reading a stream, this type-level function maps the result you will have whether you read a regular stream or $all stream. When reading a regular stream, some read-error can occur like the stream got deleted. However read-error cannot occur when reading $all stream (because $all cannot get deleted).

Equations

BatchResult EventNumber = ReadResult EventNumber StreamSlice

BatchResult Position = AllSlice

data ResolveLink

Determines whether any link event encountered in the stream will be resolved. See the discussion for more information: https://eventstore.com/docs/dotnet-api/reading-events/index.html#resolvedevent

Constructors

ResolveLink

NoResolveLink

readEvent :: Connection -> StreamName -> EventNumber -> ResolveLink -> Maybe Credentials -> IO (Async (ReadResult EventNumber ReadEvent))

Reads a single event from given stream.

readEventsBackward

Arguments

:: Connection

-> StreamId t

-> t

-> Int32
Batch size

-> ResolveLink

-> Maybe Credentials

-> IO (Async (BatchResult t))

Reads events from a stream backward.

readEventsForward

Arguments

:: Connection

-> StreamId t

-> t

-> Int32
Batch size

-> ResolveLink

-> Maybe Credentials

-> IO (Async (BatchResult t))

Reads events from a stream forward.

getStreamMetadata :: Connection -> StreamName -> Maybe Credentials -> IO (Async StreamMetadataResult)

Asynchronously gets the metadata of a stream.

Write Operations

data StreamACL

Represents an access control list for a stream.

Constructors

StreamACL

Fields

streamACLReadRoles :: !(Maybe [Text])

Roles and users permitted to read the stream.

streamACLWriteRoles :: !(Maybe [Text])

Roles and users permitted to write to the stream.

streamACLDeleteRoles :: !(Maybe [Text])

Roles and users permitted to delete to the stream.

streamACLMetaReadRoles :: !(Maybe [Text])

Roles and users permitted to read stream metadata.

streamACLMetaWriteRoles :: !(Maybe [Text])

Roles and users permitted to write stream metadata.

data StreamMetadata

Represents stream metadata with strongly typed properties for system values and a dictionary-like interface for custom values.

Constructors

StreamMetadata

Fields

streamMetadataMaxCount :: !(Maybe Int32)

The maximum number of events allowed in the stream.

streamMetadataMaxAge :: !(Maybe TimeSpan)

The maximum age of events allowed in the stream.

streamMetadataTruncateBefore :: !(Maybe Int32)

The event number from which previous events can be scavenged. This is used to implement soft-deletion of streams.

streamMetadataCacheControl :: !(Maybe TimeSpan)

The amount of time for which the stream head is cachable.

streamMetadataACL :: !(Maybe StreamACL)

The access control list for the stream.

streamMetadataCustom :: !Object

An enumerable of key-value pairs of keys to JSON text for user-provider metadata.

getCustomPropertyValue :: StreamMetadata -> Text -> Maybe Value

Gets a custom property value from metadata.

getCustomProperty :: FromJSON a => StreamMetadata -> Text -> Maybe a

Get a custom property value from metadata.

emptyStreamACL :: StreamACL

StreamACL with no role or users whatsoever.

emptyStreamMetadata :: StreamMetadata

StreamMetadata with everything set to Nothing, using emptyStreamACL and an empty Object.

deleteStream

Arguments

:: Connection

-> StreamName

-> ExpectedVersion

-> Maybe Bool
Hard delete

-> Maybe Credentials

-> IO (Async DeleteResult)

Deletes given stream.

sendEvent :: Connection -> StreamName -> ExpectedVersion -> Event -> Maybe Credentials -> IO (Async WriteResult)

Sends a single Event to given stream.

sendEvents :: Connection -> StreamName -> ExpectedVersion -> [Event] -> Maybe Credentials -> IO (Async WriteResult)

Sends a list of Event to given stream.

setStreamMetadata :: Connection -> StreamName -> ExpectedVersion -> StreamMetadata -> Maybe Credentials -> IO (Async WriteResult)

Asynchronously sets the metadata for a stream.

Builder

type Builder a = Endo a

Allows to build a structure using Monoid functions.

Stream ACL Builder

type StreamACLBuilder = Builder StreamACL

A Builder applies to StreamACL.

buildStreamACL :: StreamACLBuilder -> StreamACL

Builds a StreamACL from a StreamACLBuilder.

modifyStreamACL :: StreamACLBuilder -> StreamACL -> StreamACL

Modifies a StreamACL using a StreamACLBuilder.

setReadRoles :: [Text] -> StreamACLBuilder

Sets role names with read permission for the stream.

setReadRole :: Text -> StreamACLBuilder

Sets a single role name with read permission for the stream.

setWriteRoles :: [Text] -> StreamACLBuilder

Sets role names with write permission for the stream.

setWriteRole :: Text -> StreamACLBuilder

Sets a single role name with write permission for the stream.

setDeleteRoles :: [Text] -> StreamACLBuilder

Sets role names with delete permission for the stream.

setDeleteRole :: Text -> StreamACLBuilder

Sets a single role name with delete permission for the stream.

setMetaReadRoles :: [Text] -> StreamACLBuilder

Sets role names with metadata read permission for the stream.

setMetaReadRole :: Text -> StreamACLBuilder

Sets a single role name with metadata read permission for the stream.

setMetaWriteRoles :: [Text] -> StreamACLBuilder

Sets role names with metadata write permission for the stream.

setMetaWriteRole :: Text -> StreamACLBuilder

Sets a single role name with metadata write permission for the stream.

Stream Metadata Builder

type StreamMetadataBuilder = Builder StreamMetadata

A Builder applies to StreamMetadata.

buildStreamMetadata :: StreamMetadataBuilder -> StreamMetadata

Builds a StreamMetadata from a StreamMetadataBuilder.

modifyStreamMetadata :: StreamMetadataBuilder -> StreamMetadata -> StreamMetadata

Modifies a StreamMetadata using a StreamMetadataBuilder

setMaxCount :: Int32 -> StreamMetadataBuilder

Sets the maximum number of events allowed in the stream.

setMaxAge :: TimeSpan -> StreamMetadataBuilder

Sets the maximum age of events allowed in the stream.

setTruncateBefore :: Int32 -> StreamMetadataBuilder

Sets the event number from which previous events can be scavenged.

setCacheControl :: TimeSpan -> StreamMetadataBuilder

Sets the amount of time for which the stream head is cachable.

setACL :: StreamACL -> StreamMetadataBuilder

Overwrites any previous StreamACL by the given one in a StreamMetadataBuilder.

modifyACL :: StreamACLBuilder -> StreamMetadataBuilder

Updates a StreamMetadata's StreamACL given a StreamACLBuilder.

setCustomProperty :: ToJSON a => Text -> a -> StreamMetadataBuilder

Sets a custom metadata property.

Transaction

data Transaction

Represents a multi-request transaction with the EventStore.

data TransactionId

The id of a Transaction.

startTransaction

Arguments

:: Connection

-> StreamName
Stream name

-> ExpectedVersion

-> Maybe Credentials

-> IO (Async Transaction)

Starts a transaction on given stream.

transactionId :: Transaction -> TransactionId

Gets the id of a Transaction.

transactionCommit :: Transaction -> Maybe Credentials -> IO (Async WriteResult)

Asynchronously commits this transaction.

transactionRollback :: Transaction -> IO ()

There isn't such of thing in EventStore parlance. Basically, if you want to rollback, you just have to not transactionCommit a Transaction.

transactionWrite :: Transaction -> [Event] -> Maybe Credentials -> IO (Async ())

Asynchronously writes to a transaction in the EventStore.

Subscription

data SubscriptionClosed

Constructors

SubscriptionClosed (Maybe SubDropReason)

class SubscriptionStream s t | t -> s where

Returns the stream of a subscription.

Methods

subscriptionStream :: s -> StreamId t

class Subscription s

Common operations supported by a subscription.

Minimal complete definition

nextSubEvent, unsubscribe

data SubDropReason

Indicates why a subscription has been dropped.

Constructors

SubUnsubscribed
Subscription connection has been closed by the user.

SubAccessDenied
The current user is not allowed to operate on the supplied stream.

SubNotFound
Given stream name doesn't exist.

SubPersistDeleted
Given stream is deleted.

SubAborted
Occurs when the user shutdown the connection from the server or if the connection to the server is no longer possible.

SubNotAuthenticated (Maybe Text)

SubServerError (Maybe Text)
Unexpected error from the server.

SubNotHandled !NotHandledReason !(Maybe MasterInfo)

SubClientError !Text

SubSubscriberMaxCountReached

data SubDetails

Subscription runtime details. Not useful for the user but at least it makes Haddock documentation generation less ugly.

data SubAction

Constructors

Submit ResolvedEvent

Dropped SubDropReason

Confirmed SubDetails

unsubscribe :: Subscription s => s -> IO ()

Asynchronously unsubscribe from a subscription.

nextSubEvent :: Subscription s => s -> IO SubAction

Asks for the next subcription event. If that function is called after a SubDropped event, expect it to hang indefinitely.

streamSubEvents :: Subscription s => s -> Stream (Of SubAction) IO ()

Streams a subscription events. The stream will end when hitting Dropped event but will still emit it.

streamSubResolvedEvents :: Subscription s => s -> Stream (Of ResolvedEvent) IO ()

Like streamSubEvent but will only emit ResolvedEvent.

Volatile Subscription

data RegularSubscription t

Also referred as volatile subscription. For example, if a stream has 100 events in it when a subscriber connects, the subscriber can expect to see event number 101 onwards until the time the subscription is closed or dropped.

subscribe :: Connection -> StreamId t -> ResolveLink -> Maybe Credentials -> IO (RegularSubscription t)

Subscribes to a stream.

Catch-up Subscription

data CatchupSubscription t

This kind of subscription specifies a starting point, in the form of an event number or transaction file position. The given function will be called for events from the starting point until the end of the stream, and then for subsequently written events.

For example, if a starting point of 50 is specified when a stream has 100 events in it, the subscriber can expect to see events 51 through 100, and then any events subsequently written until such time as the subscription is dropped or closed.

subscribeFrom

Arguments

:: Connection

-> StreamId t

-> ResolveLink

-> Maybe t

-> Maybe Int32
Batch size

-> Maybe Credentials

-> IO (CatchupSubscription t)

Subscribes to a stream. If last checkpoint is defined, this will readStreamEventsForward from that event number, otherwise from the beginning. Once last stream event reached up, a subscription request will be sent using subscribe.

Persistent Subscription

data PersistentSubscription

The server remembers the state of the subscription. This allows for many different modes of operations compared to a regular or catchup subscription where the client holds the subscription state. (Need EventStore >= v3.1.0).

data PersistentSubscriptionSettings

Gathers every persistent subscription property.

Constructors

PersistentSubscriptionSettings

Fields

psSettingsResolveLinkTos :: !Bool

Whether or not the persistent subscription should resolve linkTo events to their linked events.

psSettingsStartFrom :: !Int64

Where the subscription should start from (position).

psSettingsExtraStats :: !Bool

Whether or not in depth latency statistics should be tracked on this subscription.

psSettingsMsgTimeout :: !TimeSpan

The amount of time after which a message should be considered to be timeout and retried.

psSettingsMaxRetryCount :: !Int32

The maximum number of retries (due to timeout) before a message get considered to be parked.

psSettingsLiveBufSize :: !Int32

The size of the buffer listening to live messages as they happen.

psSettingsReadBatchSize :: !Int32

The number of events read at a time when paging in history.

psSettingsHistoryBufSize :: !Int32

The number of events to cache when paging through history.

psSettingsCheckPointAfter :: !TimeSpan

The amount of time to try checkpoint after.

psSettingsMinCheckPointCount :: !Int32

The minimum number of messages to checkpoint.

psSettingsMaxCheckPointCount :: !Int32

The maximum number of message to checkpoint. If this number is reached, a checkpoint will be forced.

psSettingsMaxSubsCount :: !Int32

The maximum number of subscribers allowed.

psSettingsNamedConsumerStrategy :: !SystemConsumerStrategy

The strategy to use for distributing events to client consumers.

data SystemConsumerStrategy

System supported consumer strategies for use with persistent subscriptions.

Constructors

DispatchToSingle
Distributes events to a single client until the bufferSize is reached. After which the next client is selected in a round robin style, and the process is repeated.

RoundRobin
Distributes events to all clients evenly. If the client buffer-size is reached the client is ignored until events are acknowledged/not acknowledged.

Pinned
For use with an indexing projection such as the system $by_category projection. Event Store inspects event for its source stream id, hashing the id to one of 1024 buckets assigned to individual clients. When a client disconnects it's buckets are assigned to other clients. When a client connects, it is assigned some of the existing buckets. This naively attempts to maintain a balanced workload. The main aim of this strategy is to decrease the likelihood of concurrency and ordering issues while maintaining load balancing. This is not a guarantee, and you should handle the usual ordering and concurrency issues.

data NakAction

Gathers every possible Nak actions.

Constructors

NA_Unknown

NA_Park

NA_Retry

NA_Skip

NA_Stop

data PersistActionException

Enumerates all persistent action exceptions.

Constructors

PersistActionFail
The action failed.

PersistActionAlreadyExist
Happens when creating a persistent subscription on a stream with a group name already taken.

PersistActionDoesNotExist
An operation tried to do something on a persistent subscription or a stream that don't exist.

PersistActionAccessDenied
The current user is not allowed to operate on the supplied stream or persistent subscription.

PersistActionAborted
That action has been aborted because the user shutdown the connection to the server or the connection to the server is no longer possible.

acknowledge :: PersistentSubscription -> ResolvedEvent -> IO ()

Acknowledges that ResolvedEvent has been successfully processed.

acknowledgeEvents :: PersistentSubscription -> [ResolvedEvent] -> IO ()

Acknowledges those ResolvedEvents have been successfully processed.

failed :: PersistentSubscription -> ResolvedEvent -> NakAction -> Maybe Text -> IO ()

Mark a message that has failed processing. The server will take action based upon the action parameter.

eventsFailed :: PersistentSubscription -> [ResolvedEvent] -> NakAction -> Maybe Text -> IO ()

Mark messages that have failed processing. The server will take action based upon the action parameter.

notifyEventsProcessed :: PersistentSubscription -> [UUID] -> IO ()

Acknowledges those event ids have been successfully processed.

notifyEventsFailed :: PersistentSubscription -> NakAction -> Maybe Text -> [UUID] -> IO ()

Acknowledges those event ids have failed to be processed successfully.

defaultPersistentSubscriptionSettings :: PersistentSubscriptionSettings

System default persistent subscription settings.

createPersistentSubscription :: Connection -> Text -> StreamName -> PersistentSubscriptionSettings -> Maybe Credentials -> IO (Async (Maybe PersistActionException))

Asynchronously create a persistent subscription group on a stream.

updatePersistentSubscription :: Connection -> Text -> StreamName -> PersistentSubscriptionSettings -> Maybe Credentials -> IO (Async (Maybe PersistActionException))

Asynchronously update a persistent subscription group on a stream.

deletePersistentSubscription :: Connection -> Text -> StreamName -> Maybe Credentials -> IO (Async (Maybe PersistActionException))

Asynchronously delete a persistent subscription group on a stream.

connectToPersistentSubscription :: Connection -> Text -> StreamName -> Int32 -> Maybe Credentials -> IO PersistentSubscription

Asynchronously connect to a persistent subscription given a group on a stream.

Results

data Slice t

Gathers common slice operations.

Constructors

SliceEndOfStream

Slice ![ResolvedEvent] !(Maybe t)

sliceEvents :: Slice t -> [ResolvedEvent]

Gets slice's ResolvedEventss.

sliceEOS :: Slice t -> Bool

If the slice has reached the end of the stream.

sliceNext :: Slice t -> Maybe t

Gets the next location of this slice.

emptySlice :: Slice t

Empty slice.

type AllSlice = Slice Position

Represents a slice of the $all stream.

newtype DeleteResult

Returned after deleting a stream. Position of the write.

Constructors

DeleteResult Position

data WriteResult

Returned after writing to a stream.

Constructors

WriteResult

Fields

writeNextExpectedVersion :: !Int64

Next expected version of the stream.

writePosition :: !Position

Position of the write.

data ReadResult t a where

Enumeration detailing the possible outcomes of reading a stream.

Constructors

ReadSuccess :: a -> ReadResult t a

ReadNoStream :: ReadResult EventNumber a

ReadStreamDeleted :: StreamName -> ReadResult EventNumber a

ReadNotModified :: ReadResult t a

ReadError :: Maybe Text -> ReadResult t a

ReadAccessDenied :: StreamId t -> ReadResult t a

data RecordedEvent

Represents a previously written event.

Constructors

RecordedEvent

Fields

recordedEventStreamId :: !Text

The event stream that this event belongs to.

recordedEventId :: !UUID

Unique identifier representing this event.

recordedEventNumber :: !Int64

Number of this event in the stream.

recordedEventType :: !Text

Type of this event.

recordedEventData :: !ByteString

Representing the data of this event.

recordedEventMetadata :: !(Maybe ByteString)

Representing the metadada associated with this event.

recordedEventIsJson :: !Bool

Indicates whether the content is internally marked as json.

recordedEventCreated :: !(Maybe UTCTime)

Representing when this event was created in the system.

data ReadEvent

Represents the result of looking up a specific event number from a stream.

Constructors

ReadEventNotFound

Fields

readEventStream :: !Text

readEventNumber :: !Int64

ReadEvent

Fields

readEventStream :: !Text

readEventNumber :: !Int64

readEventResolved :: !ResolvedEvent

type StreamSlice = Slice EventNumber

Regular stream slice.

data Position

A structure referring to a potential logical record position in the EventStore transaction file.

Constructors

Position

Fields

positionCommit :: !Int64

Commit position of the record

positionPrepare :: !Int64

Prepare position of the record

data ReadDirection

Represents the direction of read operation (both from $all an usual streams).

Constructors

Forward
From beginning to end

Backward
From end to beginning

data ResolvedEvent

A structure representing a single event or an resolved link event.

Constructors

ResolvedEvent

Fields

resolvedEventRecord :: !(Maybe RecordedEvent)

The event, or the resolved link event if this ResolvedEvent is a link event.

resolvedEventLink :: !(Maybe RecordedEvent)

The link event if this ResolvedEvent is a link event.

resolvedEventPosition :: !(Maybe Position)

Possible Position of that event.

data OperationError

Operation exception that can occurs on an operation response.

Constructors

WrongExpectedVersion !Text !ExpectedVersion
Stream and Expected Version

StreamDeleted !StreamName
Stream

InvalidTransaction

forall t. AccessDenied !(StreamId t)
Stream

InvalidServerResponse !Command !Command
Expected, Found

ProtobufDecodingError !String

ServerError !(Maybe Text)
Reason

InvalidOperation !Text

StreamNotFound !StreamName

NotAuthenticatedOp
Invalid operation state. If happens, it's a driver bug.

Aborted
Occurs when the user asked to close the connection or if the connection can't reconnect anymore.

ConnectionHasDropped

data StreamId loc where

Represents a regular stream name or $all stream.

Constructors

StreamName :: Text -> StreamId EventNumber

All :: StreamId Position

type StreamName = StreamId EventNumber

isAllStream :: StreamId t -> Bool

If the stream is the $all stream.

isEventResolvedLink :: ResolvedEvent -> Bool

Indicates whether this ResolvedEvent is a resolved link event.

resolvedEventOriginal :: ResolvedEvent -> RecordedEvent

Returns the event that was read or which triggered the subscription.

If this ResolvedEvent represents a link event, the link will be the original event, otherwise it will be the event.

resolvedEventDataAsJson :: FromJSON a => ResolvedEvent -> Maybe a

Tries to desarialize resolvedEventOriginal data as JSON.

resolvedEventOriginalStreamId :: ResolvedEvent -> Text

The stream name of the original event.

resolvedEventOriginalId :: ResolvedEvent -> UUID

The ID of the original event.

resolvedEventOriginalEventNumber :: ResolvedEvent -> Int64

The event number of the original event.

recordedEventDataAsJson :: FromJSON a => RecordedEvent -> Maybe a

Tries to parse JSON object from the given RecordedEvent.

positionStart :: Position

Representing the start of the transaction file.

positionEnd :: Position

Representing the end of the transaction file.

Logging

data LogLevel

Constructors

LevelDebug

LevelInfo

LevelWarn

LevelError

LevelOther Text

type LogType = LogType' LogStr

data LoggerFilter

Constructors

LoggerFilter (LogSource -> LogLevel -> Bool)

LoggerLevel LogLevel

Monitoring

data MonitoringBackend

Monitoring backend abstraction. Gathers all the metrics currently tracked by the client. Used only by the TCP interface. Be careful as MonitoringBackend is used in a very tight loop. Each function must not throw any exception or the client will end in a broken state.

Constructors

MonitoringBackend

Fields

monitoringBackendIncrPkgCount :: IO ()

Called every time a TCP package is sent. We mean high-level TCP package, used in EventStore TCP protocol.

monitoringBackendIncrConnectionDrop :: IO ()

Called every time the client has lost the connection.

monitoringBackendAddDataTransmitted :: Int -> IO ()

When the client sends a TCP package, it calls that function by passing the size of the payload. The goal is to have a distrubtion of the amount of data exchanged with the server.

monitoringBackendIncrForceReconnect :: IO ()

Called every time the client is asked by a node to connect to another node. It happens only in cluster connection setting.

monitoringBackendIncrHeartbeatTimeouts :: IO ()

Called every time the client detects a heartbeat timeout from the server.

noopMonitoringBackend :: MonitoringBackend

A MonitoringBackend that does nothing.

Misc

data Command

Internal command representation.

data DropReason

Represents the reason subscription drop happened.

Constructors

D_Unsubscribed

D_AccessDenied

D_NotFound

D_PersistentSubscriptionDeleted

D_SubscriberMaxCountReached

data ExpectedVersion

Constants used for expected version control.

The use of expected version can be a bit tricky especially when discussing idempotency assurances given by the EventStore.

The EventStore will assure idempotency for all operations using any value in ExpectedVersion except for anyStream. When using anyStream the EventStore will do its best to assure idempotency but will not guarantee idempotency.

anyVersion :: ExpectedVersion

This write should not conflict with anything and should always succeed.

noStreamVersion :: ExpectedVersion

The stream being written to should not yet exist. If it does exist treat that as a concurrency problem.

emptyStreamVersion :: ExpectedVersion

The stream should exist and should be empty. If it does not exist or is not empty, treat that as a concurrency problem.

exactEventVersion :: Int64 -> ExpectedVersion

States that the last event written to the stream should have a sequence number matching your expected value.

streamExists :: ExpectedVersion

The stream should exist. If it or a metadata stream does not exist treat that as a concurrency problem.

msDiffTime :: Float -> NominalDiffTime

Millisecond timespan

Re-export

(<>) :: Semigroup a => a -> a -> a infixr 6

An associative operation.

data NonEmpty a

Non-empty (and non-strict) list type.

Since: base-4.9.0.0

Constructors

a :| [a] infixr 5

nonEmpty :: [a] -> Maybe (NonEmpty a)

nonEmpty efficiently turns a normal list into a NonEmpty stream, producing Nothing if the input is empty.

data TLSSettings

TLS Settings that can be either expressed as simple settings, or as full blown TLS.Params settings.

Unless you need access to parameters that are not accessible through the simple settings, you should use TLSSettingsSimple.

data NominalDiffTime

This is a length of time, as measured by UTC. It has a precision of 10^-12 s.

Conversion functions will treat it as seconds. For example, (0.010 :: NominalDiffTime) corresponds to 10 milliseconds.

It ignores leap-seconds, so it's not necessarily a fixed amount of clock time. For instance, 23:00 UTC + 2 hours of NominalDiffTime = 01:00 UTC (+ 1 day), regardless of whether a leap-second intervened.