Database.EventStore.Streaming

Description

Documentation

data ReadError t where

Constructors

StreamDeleted :: StreamName -> ReadError EventNumber

ReadError :: Maybe Text -> ReadError t

AccessDenied :: StreamId t -> ReadError t

NoStream :: ReadError EventNumber

data Fetch t

Constructors

FetchError !(ReadError t)

Fetch !(Slice t)

newtype ReadResultHandler

Constructors

ReadResultHandler

Fields

runReadResultHandler :: forall t. StreamId t -> BatchResult t -> Fetch t

readThroughForward :: Connection -> StreamId t -> ResolveLink -> t -> Maybe Int32 -> Maybe Credentials -> Stream (Of ResolvedEvent) (ExceptT (ReadError t) IO) ()

Returns an iterator able to consume a stream entirely. When reading forward, the iterator ends when the last stream's event is reached.

readThroughBackward :: Connection -> StreamId t -> ResolveLink -> t -> Maybe Int32 -> Maybe Credentials -> Stream (Of ResolvedEvent) (ExceptT (ReadError t) IO) ()

Returns an iterator able to consume a stream entirely. When reading backward, the iterator ends when the first stream's event is reached.

throwOnError :: (Show t, Typeable t) => Stream (Of a) (ExceptT (ReadError t) IO) () -> Stream (Of a) IO ()

Throws an exception in case ExceptT is a Left.

defaultReadResultHandler :: ReadResultHandler

onRegularStream :: (ReadResult EventNumber (Slice EventNumber) -> Fetch EventNumber) -> ReadResultHandler

readThrough :: Connection -> ReadResultHandler -> ReadDirection -> StreamId t -> ResolveLink -> t -> Maybe Int32 -> Maybe Credentials -> Stream (Of ResolvedEvent) (ExceptT (ReadError t) IO) ()

Returns an iterator able to consume a stream entirely.